# Speedup by reorganizing code