--- a/src/index.js
+++ b/src/index.js
@@ -1,21 +1,27 @@
 export default function cc(classes) {
   var out = ""
-  var type = typeof classes
 
-  if (type === "string" || type === "number") {
-    return classes || ""
-  }
+  if (Array.isArray(classes)) {
+    if (classes.length > 1) {
+      for (var i = 0, len = classes.length; i < len; i++) {
+        var t = typeof classes[i];
 
-  if (Array.isArray(classes) && classes.length > 0) {
-    for (var i = 0, len = classes.length; i < len; i++) {
-      var next = cc(classes[i])
-      if (next) {
-        out += (out && " ") + next
+        var next = t !== 'string' && t !== 'number' ? cc(classes[i]) : classes[i]
+        if (next) {
+          out += (out && " ") + next
+        }
       }
     }
+
   } else {
+    var type = typeof classes
+
+    if (type === "string" || type === "number") {
+      return classes || ""
+    }
+
     for (var key in classes) {
-      if (classes.hasOwnProperty(key) && classes[key]) {
+      if (classes[key]) {
         out += (out && " ") + key
       }
     }