By reordering some of the logic, we can optimize the function's runtime efficiency.

By doing the Array.isArray() check first, we catch the most common cases and can elide the (moderately expensive) typeof operations.  By separating the isArray() check from the length check, we can elide even more operations in the case where an empty array is passed.  We can reduce even more operations by not recursing when there are simple (string/number) types in the array and concatenating them directly.

This fork should be faster in all cases.

Here's some benchmark results (classcat = my fork, classcatNpm = as it exists now).  

```
Classcat – Strings × 9,316,220 ops/sec
ClasscatNPM – Strings × 7,874,574 ops/sec
classNames – Strings × 3,094,206 ops/sec
Fastest is Classcat – Strings

Classcat – Objects × 95,576,430 ops/sec
ClasscatNPM – Objects × 6,921,587 ops/sec
classNames – Objects × 2,897,037 ops/sec
Fastest is Classcat – Objects

Classcat – Strings & Objects × 5,738,749 ops/sec
ClasscatNPM – Strings & Objects × 5,516,302 ops/sec
classNames – Strings & Objects × 2,589,446 ops/sec
Fastest is Classcat – Strings & Objects

Classcat – Mixed × 3,374,364 ops/sec
ClasscatNPM – Mixed × 3,259,550 ops/sec
classNames – Mixed × 1,857,683 ops/sec
Fastest is Classcat – Mixed

Classcat – Arrays × 3,121,509 ops/sec
ClasscatNPM – Arrays × 2,603,359 ops/sec
classNames – Arrays × 780,863 ops/sec
Fastest is Classcat – Arrays
```