@jorgebucaran 2018-05-16
@EdwardDrapkin The length > 1 check makes this particular case fail:

```js
cc(["ok"])
```

...which should be infrequent, but still surprising.