@EdwardDrapkin 2018-05-16
Alright, I updated the check to be length > 0 (whoops lol) and I also added a test case for it.

I also optimized the for loop based on this blog post: https://medium.com/@AamuLumi/voyage-to-the-most-efficient-loop-in-nodejs-and-a-bit-js-5961d4524c2e 

I can't notice any improvement, but I'm testing on a laptop so it's probably being lost in statistical noise.

I also changed the semantics a little bit in that passing 0 will result in '0' instead of '' because I figured if someone is passing a number, they are doing some kind of weird class replacement based on array indices in a babel plugin or something and they would want their 0 to be preserved.