@EdwardDrapkin 2018-05-16
A minor beneficial side effect is that you can now pass nested arrays (added a test case): `[foo, [bar, [foo-bar]], bar-foo, [[[[test]]]]]` winds up becoming "foo bar foo-bar bar-foo test".