@jorgebucaran 2018-05-16
@EdwardDrapkin Ah, the numbers as CSS classes debate. 😄

Technically a class that starts with a number is not valid CSS. But even if it was (or is going to be in some future spec), a number on its own is semantically meaningless without some sort of prefix, so I'd prefer to optimize for size in this instance.

> I can't notice any improvement, but I'm testing on a laptop so it's probably being lost in statistical noise.

Haha bummer. I was quite surprised when I saw that 95,576,430 ops/sec at first.