@jorgebucaran 2018-08-18
> @EdwardDrapkin A minor beneficial side effect is that you can now pass nested arrays (added a test case): `[foo, [bar, [foo-bar]], bar-foo, [[[[test]]]]]` winds up becoming "foo bar foo-bar bar-foo test".

This can be done with our current version of classcat. I added a test for that in 85af3f5.

>  also changed the semantics a little bit in that passing 0 will result in '0' instead of ''...

I am going to pass on this one for now. `0` is still converted to `""`, so we'll have to cast to string the value if there's a chance the number 0 could be used as a class name.

---

Closing as there was no noticeable perf result difference and for the reasons given above. 

Thank you, anyway @EdwardDrapkin! 👋